feat #9 :truck: move to ui package
