+++ presentation/src/main/java/com/dcs/presentation/ui/main/MainBottomNavigation.kt
@@ -1,4 +1,4 @@
-package com.dcs.presentation.main
+package com.dcs.presentation.ui.main
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.material3.Icon
@@ -16,7 +16,7 @@ import androidx.navigation.NavDestination.Companion.hierarchy
 import androidx.navigation.NavGraph.Companion.findStartDestination
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.currentBackStackEntryAsState
-import com.dcs.presentation.main.Screen.Main
+import com.dcs.presentation.ui.main.Screen.Main
 
 @Composable
 fun MainBottomNavigation(
