+++ presentation/src/main/java/com/dcs/presentation/ui/people/PeopleScreen.kt
@@ -1,4 +1,4 @@
-package com.dcs.presentation.people
+package com.dcs.presentation.ui.people
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
