+++ presentation/src/main/java/com/dcs/presentation/ui/setting/SettingEffect.kt
@@ -1,4 +1,4 @@
-package com.dcs.presentation.setting
+package com.dcs.presentation.ui.setting
 
 sealed interface SettingEffect {
     data object SignIn : SettingEffect
