+++ presentation/src/main/java/com/dcs/presentation/ui/setting/SettingEvent.kt
@@ -1,4 +1,4 @@
-package com.dcs.presentation.setting
+package com.dcs.presentation.ui.setting
 
 sealed interface SettingEvent {
     data object SignIn : SettingEvent
